[App] split the cook driver and the ui out of AppFramework
AppFramework.cpp held three unrelated jobs in 1135 lines: the application
lifecycle, the whole build-time cook driver, and the imgui control panel. The
cook half is a sibling of Cooker and SceneCooker, not of the app framework, and
being file-local static made its job planning and product manifest unreachable
from tests.

RunCookPipeline in CookPipeline.cpp now owns the plan, the HDR transcode
derivation and the products manifest, taking the RHI accelerator as a parameter
rather than reaching for a member. AppFramework::RunCookMode keeps only what it
owns: bringing up the optional RHI, resolving the targets, and tearing the core
down. The control panel moves to AppFrameworkUi.cpp. AppFramework.cpp is 714
lines and 20 includes lighter.

cook_targets covers the target table and the '+'-separated grammar, which had no
test while it was static in AppFramework: an unknown target must reject the whole
list rather than cook a subset, and this build's own target must be in the table.
