+++ docs/Cooking.md
@@ -20,7 +20,7 @@ The cook stage's product is one self-contained content image per cook target: ev
 2. Request it at the natural trigger (attach, resource init) with `Cooker::Request`. Prefer the key + factory overload: a cache hit then never loads or decodes the source. The payload arrives on the main thread.
 3. Give the requester a valid pending state — rendering must work before delivery. The sky light renders a flat sky until its cube map arrives.
 4. Register the request as a scene async task so screenshot tests and USD export wait for the cook; `SkyLight::RequestCook` is the reference for the whole request-deliver-apply pattern.
-5. Wire the job into the application's cook plan (`AppFramework::RunCookMode`) so release cooks produce it, and extend [cook_list.json](../resources/packed/config/cook_list.json) if it depends on a scene not cooked yet.
+5. Wire the job into the build-time cook plan (`RunCookPipeline` in [libraries/source/application/CookPipeline.cpp](../libraries/source/application/CookPipeline.cpp)) so release cooks produce it, and extend [cook_list.json](../resources/packed/config/cook_list.json) if it depends on a scene not cooked yet.
 6. Optional GPU acceleration must produce byte-identical artifacts under the same key — see CPU/GPU parity below.
 
 ## Artifacts
@@ -51,7 +51,7 @@ Unlike the async SkyLight pattern, texture cooks resolve synchronously inside th
 
 The compressed image stays a regular `Image2D`. Consumers that need texels — the CPU pipeline's material sampling and USD export — decode lazily to an RGBA8 shadow copy on first use. If a device cannot sample the format (`RHIContext::SupportsSampledFormat`), `CreateTexture` uploads that decoded copy instead; software rasterizers (lavapipe, SwiftShader) sample both families directly.
 
-Only the requested cook targets' families cook (the target→family mapping is `CookTargetFamilies` in AppFramework.cpp); a build-time cook writes `cooked/cook_products.json`, mapping each target to its plan's manifest keys and to the source images those artifacts fully replace. Image assembly (`assemble_cooked_image` in build.py) projects the pool through that file: a target's image carries exactly its plan's artifacts — never the other family — and a replaced source is left out only when every replacing artifact verifies (an unverifiable source ships and stays the runtime fallback). Packaging ([dev/package_cooked.py](../dev/package_cooked.py)) swaps in the product's own image and asserts its self-declared target instead of carrying any content policy. Inside a stripped package the image load fails, tydra keeps the authored path on the image entry, and the slot resolves from the manifest by identity alone — with no source pixels an inline cook is impossible, so a missing artifact degrades to the dummy texture instead of poisoning the cache. `rebuild_cache true` in a stripped package cannot re-encode textures — it keeps using the packaged artifacts.
+Only the requested cook targets' families cook (the target→family mapping is [cook_targets.json](../cook_targets.json), which CMake compiles into `CookTargets` and both build.py and dev/ci_matrix.py read); a build-time cook writes `cooked/cook_products.json`, mapping each target to its plan's manifest keys and to the source images those artifacts fully replace. Image assembly (`assemble_cooked_image` in build.py) projects the pool through that file: a target's image carries exactly its plan's artifacts — never the other family — and a replaced source is left out only when every replacing artifact verifies (an unverifiable source ships and stays the runtime fallback). Packaging ([dev/package_cooked.py](../dev/package_cooked.py)) swaps in the product's own image and asserts its self-declared target instead of carrying any content policy. Inside a stripped package the image load fails, tydra keeps the authored path on the image entry, and the slot resolves from the manifest by identity alone — with no source pixels an inline cook is impossible, so a missing artifact degrades to the dummy texture instead of poisoning the cache. `rebuild_cache true` in a stripped package cannot re-encode textures — it keeps using the packaged artifacts.
 
 The `texture_compression` test case covers encode/decode invariants for every profile and family on every platform; packaged screenshot runs plus the CI cook gate prove end to end that shipped packages resolve every texture from artifacts.
